New version of generate_post_file().
Now elegantly handles the situation when the file already
exists.

Mode while opening the file is now 'x'. This means, creation of
the file fails if the file already exists.

Also, filename is no longer passed into the function. Instead,
it is determined from within the function as is passed out as
the return-value.